Auto-Refresh the log page after submitting a user action
