+++ src/analytics/UserInfo.tsx
@@ -1,10 +1,6 @@
 import React, { useEffect, useState } from 'react';
 
 import { makeStyles } from '@material-ui/core/styles';
-import { Button } from '@material-ui/core';
-import { useSelector } from 'react-redux';
-import { getUserGroups } from '../auth';
-import { userActions } from '../utils/api';
 
 const useStyles = makeStyles((theme) => ({
   wrapper: {
@@ -78,21 +74,10 @@ interface UserInfoProps {
 
 const UserInfo = function ({ userData }: UserInfoProps) {
   const classes = useStyles();
-  const userGroups = useSelector(getUserGroups);
   const [selectedQuestionnaire, setSelectedQuestionnaire] = useState('');
-  const [action, setAction] = useState('');
-  const [collection, setCollection] = useState('');
-  const [module, setModule] = useState('');
-  const [questionnaire, setQuestionnaire] = useState('');
-  const [actionError, setActionError] = useState('');
 
   useEffect(() => {
     setSelectedQuestionnaire('');
-    setAction('');
-    setCollection('');
-    setModule('');
-    setQuestionnaire('');
-    setActionError('');
   }, [userData]);
 
   const handleQuestionnaire = (e: {
@@ -103,234 +88,8 @@ const UserInfo = function ({ userData }: UserInfoProps) {
     setSelectedQuestionnaire(e.target.value);
   };
 
-  const handleUserAction = (event: { preventDefault: () => void }) => {
-    event.preventDefault();
-    let username = userData['username'];
-    let action_content = '';
-    if (action === 'add_module') {
-      action_content = module;
-    } else if (action === 'remove_module') {
-      action_content = module;
-    } else if (action === 'complete_questionnaire') {
-      action_content = questionnaire;
-    } else if (action === 'delete_user') {
-      action_content = username;
-    } else if (action === 'update_recipe_collection') {
-      action_content = collection;
-    }
-
-    if (action_content !== '') {
-      userActions(username, action, action_content)
-        .then((res: any) => {
-          if (res.data.status === true) {
-            setActionError('User action performed successfully.');
-          } else {
-            setActionError('User action failed.');
-          }
-        })
-        .catch((error) => {
-          setActionError(error.response.data.detail);
-        });
-    } else {
-      setActionError('No action item selected');
-    }
-  };
-
   return (
     <div className={classes.wrapper}>
-      {userGroups.includes('admin_user_analytics') ||
-      userGroups.includes('care_panel_admin_analytics') ? (
-        <div>
-          <strong>User Action: </strong>
-          <select
-            onChange={(e) => {
-              setAction(e.target.value);
-            }}
-            value={action}
-            className={classes.select}
-          >
-            <option value="" className={classes.selectOption}>
-              Select User Action
-            </option>
-            <option
-              key="add_module"
-              value="add_module"
-              className={classes.selectOption}
-            >
-              Add Module
-            </option>
-            <option
-              key="remove_module"
-              value="remove_module"
-              className={classes.selectOption}
-            >
-              Remove Module
-            </option>
-            <option
-              key="complete_questionnaire"
-              value="complete_questionnaire"
-              className={classes.selectOption}
-            >
-              Complete Questionnaire
-            </option>
-            <option
-              key="delete_user"
-              value="delete_user"
-              className={classes.selectOption}
-            >
-              Delete Account
-            </option>
-            <option
-              key="update_recipe_collection"
-              value="update_recipe_collection"
-              className={classes.selectOption}
-            >
-              Update Recipe Collection
-            </option>
-          </select>
-
-          {action === 'complete_questionnaire' ? (
-            <select
-              onChange={(e) => {
-                setQuestionnaire(e.target.value);
-              }}
-              value={questionnaire}
-              className={classes.select}
-            >
-              <option value="" className={classes.selectOption}>
-                Select Questionnaire
-              </option>
-              {Object.entries(userData['questionnaires']).map(([slug]) => (
-                <option
-                  key={slug}
-                  value={slug}
-                  className={classes.selectOption}
-                >
-                  {slug}
-                </option>
-              ))}
-            </select>
-          ) : (
-            ''
-          )}
-
-          {action === 'update_recipe_collection' ? (
-            <select
-              onChange={(e) => {
-                setCollection(e.target.value);
-              }}
-              value={collection}
-              className={classes.select}
-            >
-              <option value="" className={classes.selectOption}>
-                Select Recipe Collection
-              </option>
-              <option key="1" value="1" className={classes.selectOption}>
-                Gut-Friendly
-              </option>
-              <option key="2" value="2" className={classes.selectOption}>
-                HB-Friendly
-              </option>
-              <option key="3" value="3" className={classes.selectOption}>
-                Low-FODMAP
-              </option>
-              <option key="4" value="4" className={classes.selectOption}>
-                Mediterranean-Diet
-              </option>
-              <option key="5" value="5" className={classes.selectOption}>
-                IBD-Flareup
-              </option>
-              <option key="6" value="6" className={classes.selectOption}>
-                HB-Weightloss
-              </option>
-            </select>
-          ) : (
-            ''
-          )}
-
-          {action === 'add_module' || action === 'remove_module' ? (
-            <select
-              onChange={(e) => {
-                setModule(e.target.value);
-              }}
-              value={module}
-              className={classes.select}
-            >
-              <option value="" className={classes.selectOption}>
-                Select Module
-              </option>
-              <option key="2" value="2" className={classes.selectOption}>
-                hypno
-              </option>
-              <option key="3" value="3" className={classes.selectOption}>
-                cbt
-              </option>
-              <option key="1" value="1" className={classes.selectOption}>
-                low-fodmap
-              </option>
-              <option key="21" value="21" className={classes.selectOption}>
-                HB-weight-loss
-              </option>
-              <option key="22" value="22" className={classes.selectOption}>
-                HB-weight-loss-hypnosis
-              </option>
-              <option key="14" value="14" className={classes.selectOption}>
-                HB-Nutrition
-              </option>
-              <option key="18" value="18" className={classes.selectOption}>
-                HB-Hypnosis-FD
-              </option>
-              <option key="17" value="17" className={classes.selectOption}>
-                HB-Hypnosis-GERD
-              </option>
-              <option key="13" value="13" className={classes.selectOption}>
-                HB-CBT
-              </option>
-              <option key="15" value="15" className={classes.selectOption}>
-                HB-Relaxation
-              </option>
-              <option key="19" value="19" className={classes.selectOption}>
-                IBD-PE
-              </option>
-              <option key="7" value="7" className={classes.selectOption}>
-                IBD-Low-FODMAP
-              </option>
-              <option key="11" value="11" className={classes.selectOption}>
-                IBD-Relaxation
-              </option>
-              <option key="10" value="10" className={classes.selectOption}>
-                IBD-CBT
-              </option>
-              <option key="8" value="8" className={classes.selectOption}>
-                IBD-MedDiet
-              </option>
-            </select>
-          ) : (
-            ''
-          )}
-
-          {action !== '' ? (
-            <Button
-              color="primary"
-              type="button"
-              variant="contained"
-              onClick={handleUserAction}
-            >
-              Submit
-            </Button>
-          ) : (
-            ''
-          )}
-        </div>
-      ) : (
-        ''
-      )}
-      {actionError && (
-        <p className={classes.error}>
-          <i>{actionError}</i>
-        </p>
-      )}
-      <br />
       <div className={classes.infoItem}>
         <strong>Username: </strong>
         {userData['username']}
